Create build workflow
Create a workflow file for building the application and deploy the artifacts to GitHub.
This will create 3 artifacts (kuiper32/kuiper64/ubuntu24) for better integration.
All the artifacts are Debian packages for the specific architecture.
The build is run inside containers.
For testing, the Debian package is installed and files are verified.
---------

Signed-off-by: Nicu Siderias <nicu.siderias@analog.com>